Split test methods out of the Element class.
