+++ client/element.js
@@ -1,5 +1,4 @@
 import elementData from './element-data.js';
-import Test from './test.js';
 
 /**
  * Provide information about atomic elements.
@@ -193,25 +192,6 @@ export default class Element {
         }
 
         return elements;
-    }
-
-    /**
-     * Test the find method.
-     *
-     * @returns {integer} How many tests failed
-     */
-    static findTest(){
-        const tests = [
-            [[''], new Set([])],
-            [[' '], new Set([])],
-            [['x'], new Set([])],
-            [['h'], new Set([1])],
-            [['he'], new Set([2])],
-            [['heli'], new Set([2])],
-            [['bor'], new Set([5, 106])],
-        ];
-
-        return Test.run(Element.find, tests);
     }
 
     /**
